Original change: https://android-review.googlesource.com/c/platform/system/core/+/2879846

+++ fs_mgr/libsnapshot/include/libsnapshot/cow_writer.h
@@ -59,7 +59,7 @@ struct CowOptions {
     bool batch_write = false;
 
     // Size of the cow operation buffer; used in v3 only.
-    uint32_t op_count_max = 0;
+    uint64_t op_count_max = 0;
 };
 
 // Interface for writing to a snapuserd COW. All operations are ordered; merges
